src/main/java/io/krukm/model/CoinReserve.java
package io.krukm.model;

import java.util.Stack;

public class CoinReserve {

    Stack<Coin> nickelStack = new Stack<>();
    Stack<Coin> dimeStack = new Stack<>();
    Stack<Coin> quarterStack = new Stack<>();

    public boolean coinAccepted(Coin coin) {

        switch (coin) {
            case PENNY:
                return false;
            case NICKEL:
                nickelStack.push(Coin.NICKEL);
                return true;
            case DIME:
                dimeStack.push(Coin.DIME);
                return true;
            case QUARTER:
                quarterStack.push(Coin.QUARTER);
                return true;
        }
        if (coinAccepted(coin)) return true;
        else return false;
    }

    public void stockReserve() {

        for (int i = 0; i < 10; i++) {

            nickelStack.push(Coin.NICKEL);
            dimeStack.push(Coin.DIME);
            quarterStack.push(Coin.QUARTER);
        }
    }
}
